# src/routes/tickets.js
const express = require("express");
const {
  createTicket,
  listTickets,
  getTicketById,
  updateTicketFields,
  addToProcessingLog,
} = require("../store/ticketStore");
const {
  validateCreateTicketBody,
  validateListFilters,
} = require("../validation");
const { ApiError } = require("../errors");
const { triageTicket } = require("../../skills/triage");
const { researchTicket } = require("../../subagents/researcher");
const { shouldEscalateIfNeeded, escalateTicket } = require("../../subagents/escalation");
const { prePiiHook } = require("../../hooks/pre-pii");
const { postLoggerHook, getLogsByTicketId } = require("../../hooks/post-logger");

const router = express.Router();

// Automation pipeline - processes a ticket through triage, research, etc.
const processTicketThroughPipeline = async (ticketId) => {
  try {
    // Step 1: Get the ticket
    const ticket = getTicketById(ticketId);
    if (!ticket) {
      throw new ApiError(404, "Ticket not found");
    }

    // Step 2: Apply PII pre-hook (redact sensitive information)
    const { redactedDescription, piiDetected, matches } = await prePiiHook(ticket);
    if (piiDetected) {
      await updateTicketFields(ticketId, {
        description: redactedDescription,
        pii_redacted: true
      });
      addToProcessingLog(ticketId, {
        event: "pii_redaction",
        description: "PII detected and redacted from ticket description"
      });
    }

    // Step 3: Ticket triage - classify by urgency and topic
    const descriptionToUse = ticket.pii_redacted ? ticket.description : ticket.description;
    const triageResult = await triageTicket({
      subject: ticket.subject,
      description: descriptionToUse,
      submitter_ref: ticket.submitter_ref
    });

    await updateTicketFields(ticketId, {
      urgency: triageResult.urgency,
      category: triageResult.category,
      triage_result: JSON.stringify(triageResult)
    });

    addToProcessingLog(ticketId, {
      event: "triage_completed",
      triage_result: triageResult
    });

    // Step 4: Research - look up documentation and produce draft response
    const researchResult = await researchTicket({
      subject: ticket.subject,
      description: descriptionToUse,
      category: ticket.category
    });

    await updateTicketFields(ticketId, {
      draft_response: researchResult.draft_response,
      confidence_score: researchResult.confidence
    });

    addToProcessingLog(ticketId, {
      event: "research_completed",
      confidence: researchResult.confidence,
      sources_found: researchResult.sources_found,
      draft_response_length: researchResult.draft_response ? researchResult.draft_response.length : 0
    });

    // Step 5: Check if escalation is needed
    const shouldEscalate = shouldEscalateIfNeeded({
      confidence: researchResult.confidence,
      triageResult: triageResult,
      ticketId: ticketId
    });

    if (shouldEscalate) {
      const escalationResult = await escalateTicket({
        ticketId: ticketId,
        reason: "Low confidence score or complex issue requiring human intervention",
        context: {
          ticketInfo: {
            subject: ticket.subject,
            description: ticket.description,
            submitter_ref: ticket.submitter_ref,
            status: ticket.status
          },
          triageResult: triageResult,
          researchResult: researchResult,
          processingLog: await getLogsByTicketId(ticketId)
        }
      });

      await updateTicketFields(ticketId, {
        status: "escalated",
        escalation_reason: escalationResult.escalation_reason
      });

      addToProcessingLog(ticketId, {
        event: "ticket_escalated",
        reason: escalationResult.escalation_reason
      });

      // Apply response post-hook for the draft response (even though escalated)
      await postLoggerHook({
        ticket_id: ticketId,
        response_reference: researchResult.draft_response,
        confidence_score: researchResult.confidence,
        response_type: "draft"
      });

      return; // Don't proceed to final response if escalated
    }

    // Step 6: Finalize response (use draft as final if confidence is high enough)
    const finalResponse = researchResult.confidence >= 0.7
      ? researchResult.draft_response
      : "Please note: This response requires human review due to low confidence. A support agent will review and respond shortly.";

    await updateTicketFields(ticketId, {
      final_response: finalResponse,
      status: researchResult.confidence >= 0.7 ? "closed" : "in_progress"
    });

    addToProcessingLog(ticketId, {
      event: "response_finalized",
      confidence: researchResult.confidence,
      is_auto_resolved: researchResult.confidence >= 0.7,
      final_response_length: finalResponse.length
    });

    // Step 7: Apply response post-hook (logging)
    await postLoggerHook({
      ticket_id: ticketId,
      response_reference: finalResponse,
      confidence_score: researchResult.confidence,
      response_type: "final"
    });

  } catch (error) {
    // Log error to ticket processing log
    const ticket = getTicketById(ticketId);
    if (ticket) {
      addToProcessingLog(ticketId, {
        event: "pipeline_error",
        error: error.message,
        stack: process.env.NODE_ENV === 'development' ? error.stack : undefined
      });
    }
    throw error;
  }
};

router.post("/", (req, res, next) => {
  try {
    validateCreateTicketBody(req.body);
    const ticket = createTicket(req.body);

    // Process the ticket through the automation pipeline (fire and forget)
    // In a production system, this would be queued for background processing
    processTicketThroughPipeline(ticket.ticket_id).catch(console.error);

    res.status(201).json(ticket);
  } catch (error) {
    next(error);
  }
});

router.get("/", (req, res, next) => {
  try {
    validateListFilters(req.query);
    const tickets = listTickets({
      status: req.query.status,
      urgency: req.query.urgency,
      category: req.query.category,
      search: req.query.search,
    });
    res.json(tickets);
  } catch (error) {
    next(error);
  }
});

router.get("/:id", (req, res, next) => {
  try {
    const ticket = getTicketById(req.params.id);
    if (!ticket) {
      throw new ApiError(404, "Ticket not found");
    }

    res.json(ticket);
  } catch (error) {
    next(error);
  }
});

router.patch("/:id", (req, res, next) => {
  try {
    const existingTicket = getTicketById(req.params.id);
    if (!existingTicket) {
      throw new ApiError(404, "Ticket not found");
    }

    // Allow updating any fields except system fields
    const allowedFields = ['subject', 'description', 'status', 'urgency', 'category'];
    const updates = {};
    
    for (const field of allowedFields) {
      if (req.body.hasOwnProperty(field)) {
        updates[field] = req.body[field];
      }
    }

    if (Object.keys(updates).length === 0) {
      throw new ApiError(400, "No valid fields provided for update");
    }

    const updatedTicket = updateTicketFields(req.params.id, updates);
    res.json(updatedTicket);
  } catch (error) {
    next(error);
  }
});

// DELETE endpoint to close/delete a ticket
router.delete("/:id", async (req, res, next) => {
  try {
    const ticket = getTicketById(req.params.id);
    if (!ticket) {
      throw new ApiError(404, "Ticket not found");
    }

    // Mark ticket as closed/deleted by updating status
    const closedTicket = updateTicketFields(req.params.id, {
      status: "closed",
      resolved_at: new Date().toISOString()
    });

    // Log the ticket closed event (synchronous)
    addToProcessingLog(req.params.id, {
      event: "ticket_closed",
      description: "Ticket was closed/deleted"
    });

    res.json({
      success: true,
      message: "Ticket closed successfully",
      ticket: closedTicket
    });
  } catch (error) {
    next(error);
  }
});

module.exports = router;